Pull out more util functions
